requirements.txt
pycairo~=1.27.0